+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Feat(mobile): Add in-app survey feature. The coordinator can trigger surveys which will be shown in the app.
 - Feat(webapp): show tradingview chart
 - Refactor(webapp): Replace axum-server with low lever rust-tls implementation
+- Feat(coordinator): Add endpoint to get a leadership board
 
 ## [1.8.5] - 2024-02-05